moved components folder to base directory
