+++ frontend/app/page.tsx
@@ -1,7 +1,7 @@
 import { Inter } from "next/font/google";
 import Image from "next/image";
+import Header from "../components/header/header";
 import styles from "./page.module.css";
-import Header from "./components/header/header";
 
 const inter = Inter({ subsets: ["latin"] });
